+++ DiplomaApp.Repositories/Repositories/CategoryRepository.cs
@@ -36,7 +36,7 @@ public class CategoryRepository:IRepository<Category>
         throw new NotImplementedException();
     }
 
-    public Category Read(ushort id)
+    public Category Read(int id)
     {
         throw new NotImplementedException();
     }
